// sourcecode/BTrak.Packages/snova-employeelist/module/module-components/src/lib/employeeList/models/custom-form-field.model.ts
export class CustomFormFieldModel {
    customFieldId: string;
    formName: string;
    formJson: string;
    formDataJson: string;
    createdDateTime: any;
    CreatedByUserId: string;
    UpdatedDateTime: Date;
    UpdatedByUserId: string;
    IsArchived: boolean;
    archivedDateTime: Date;
    archivedByUserId: string;
    isDeleted: boolean;
    timeStamp: any;
    formKeys: string;
    moduleTypeId: number;
    referenceId: string;
    referenceTypeId: string;
    fieldName: string;
    customFieldTimeStamp: any;
    customDataFormFieldId: string;
    formData: any;
    formCreatedDateTime: any;
}